requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==2.0.12
class-registry==2.1.2
Django==3.2
django-filter==23.4
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
filters==1.3.2
flake8==6.1.0
idna==3.6
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.2
pluggy==0.13.1
py==1.11.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2023.3.post1
regex==2023.10.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.8.0
urllib3==1.26.18
